+++ src/util/mod.rs
@@ -1,149 +1,46 @@
 //! Utility functions for unit conversion, input validations, etc.
 
 pub mod units;
-pub mod validation;
-use crate::Error;
 pub use units::*;
 
-/// Converts given amount with given unit to Qa.
-/// # Example
-/// ```
-/// use zilliqa_rs::util::parse_units;
-/// let amount_in_zil = 15230001000000_u128;
-/// let amount_in_li = 15230001_u128;
-/// let amount_in_qa = 15_u128;
-/// assert_eq!(amount_in_zil, parse_units("15.230001000000000000", "zil").unwrap());
-/// assert_eq!(amount_in_li, parse_units("15.230001000000000000", "li").unwrap());
-/// assert_eq!(amount_in_qa, parse_units("15.230001000000000000", "qa").unwrap());
-/// ```
-pub fn parse_units<K, S>(amount: S, unit: K) -> Result<u128, Error>
-where
-    S: ToString,
-    K: TryInto<Units, Error = Error> + Copy,
-{
-    let exponent: u32 = unit.try_into()?.as_num();
-    let mut amount_str = amount.to_string().replace('_', "");
-    let negative = amount_str.chars().next().unwrap_or_default() == '-';
-    let dec_len = if let Some(di) = amount_str.find('.') {
-        amount_str.remove(di);
-        amount_str[di..].len() as u32
-    } else {
-        0
-    };
-
-    if dec_len > exponent {
-        // Truncate the decimal part if it is longer than the exponent
-        let amount_str = &amount_str[..(amount_str.len() - (dec_len - exponent) as usize)];
-        if negative {
-            Err(Error::NegativeValueNotAllowed)
-        } else {
-            Ok(amount_str.parse::<u128>()?)
-        }
-    } else if negative {
-        Err(Error::NegativeValueNotAllowed)
-    } else {
-        let mut a_uint: u128 = amount_str.parse()?;
-        a_uint *= 10_u128.checked_pow(exponent - dec_len).ok_or(Error::ParseOverflow)?;
-        Ok(a_uint)
-    }
-}
-
-/// Converts given amount in Qa to requested unit.
-/// # Example
-///
-///```
-/// use zilliqa_rs::util::format_units;
-///
-///let zil = format_units(1395633240123_u128, "zil").unwrap();
-///assert_eq!(zil, "1.395633240123");
-///
-///let li = format_units(1395633240123_u128, "li").unwrap();
-///assert_eq!(li, "1395633.240123");
-///```
-///
-pub fn format_units<K>(amount: u128, unit: K) -> Result<String, Error>
-where
-    K: TryInto<Units, Error = Error>,
-{
-    let units = unit.try_into()?.as_num();
-    let exp10 = 10_u128.pow(units);
-    let integer = amount / exp10;
-    let decimals = (amount % exp10).to_string();
-    let units = units as usize;
-    Ok(format!("{integer}.{decimals:0>units$}"))
-}
-
-/// Formats a given amount in Qa into Zil.
-///
-/// # Example
-/// ```
-/// use zilliqa_rs::util::format_zil;
-///
-/// let zil = format_zil(1395633240123_u128);
-/// assert_eq!(zil, "1.395633240123");
-/// ```
-pub fn format_zil(amount: u128) -> String {
-    // Safe to call unwrap, "zil" can be converted to unit
-    format_units(amount, "zil").unwrap()
-}
-
-/// Parses an amount in ZIL into a QA.
-///
-/// # Example
-/// ```
-/// use zilliqa_rs::util::parse_zil;
-///
-/// let amount_in_qa = 15230001000000;
-/// assert_eq!(amount_in_qa, parse_zil("15.230001").unwrap());
-/// ```
-pub fn parse_zil<S: ToString>(zil: S) -> Result<u128, Error> {
-    parse_units(zil, "zil")
+/// Checks if a given string is a valid byte string of a specified length.
+pub fn is_byte_string(str: &str, len: usize) -> bool {
+    let regex = regex::Regex::new(&format!("^[0-9a-fA-F]{{{}}}$", len)).expect("Failed to create the regex for `is_byte_string`");
+    let str = str.replace("0x", "");
+    regex.is_match(&str)
 }
 
 #[cfg(test)]
 mod tests {
-    use claim::assert_err;
-
-    use super::{format_units, format_zil, parse_units, parse_zil};
+    use super::is_byte_string;
 
     #[test]
-    fn parse_units_should_work() {
-        let amount_in_zil = 15230001000000_u128;
-        let amount_in_li = 15230001_u128;
-        let amount_in_qa = 15_u128;
-        assert_eq!(amount_in_zil, parse_units("15.230001000000000000", "zil").unwrap());
-        assert_eq!(amount_in_li, parse_units("15.230001000000000000", "li").unwrap());
-        assert_eq!(amount_in_qa, parse_units("15.230001000000000000", "qa").unwrap());
+    fn is_byte_string_should_return_true_for_a_valid_byte_string_with_correct_size() {
+        let str = "1234567890";
+        assert!(is_byte_string(str, str.len()))
     }
 
     #[test]
-    fn parse_units_should_return_error_for_negative_values() {
-        assert_err!(parse_units("-15.230001000000000000", "zil"));
+    fn is_byte_string_should_return_true_for_a_valid_byte_string_with_correct_size_even_if_its_prepended_with_0x() {
+        let str = "0x1234567890";
+        assert!(is_byte_string(str, str.len() - 2)) // -2 for 0x
     }
 
     #[test]
-    fn format_units_should_work() {
-        let zil = format_units(1395633240123_u128, "zil").unwrap();
-        assert_eq!(zil, "1.395633240123");
-
-        let li = format_units(1395633240123_u128, "li").unwrap();
-        assert_eq!(li, "1395633.240123");
+    fn is_byte_string_should_return_true_for_a_valid_byte_string_with_correct_size_when_it_contains_letters_a_f() {
+        let str = "1234567890aabbccddeeff";
+        assert!(is_byte_string(str, str.len()))
     }
 
     #[test]
-    fn parse_zil_should_work() {
-        let amount_in_zil = 15230001000000;
-        assert_eq!(amount_in_zil, parse_zil("15.230001000000000000").unwrap());
+    fn is_byte_string_should_return_false_if_size_is_incorrect() {
+        let str = "1234567890aabbccddeeff";
+        assert_eq!(is_byte_string(str, str.len() - 2), false);
     }
 
     #[test]
-    fn parse_zil_should_error_for_negative_values() {
-        assert_err!(parse_zil("-15.230001000000000000"));
-    }
-
-    #[test]
-    fn format_zil_should_work() {
-        let zil = format_zil(1395633240123_u128);
-        assert_eq!(zil, "1.395633240123");
+    fn is_byte_string_should_return_false_if_contains_out_of_a_f_characters() {
+        let str = "1234567890aabbccddeeffgg";
+        assert_eq!(is_byte_string(str, str.len()), false);
     }
 }
